+++ frontend/src/app/connectionContext.ts
@@ -1,0 +1,4 @@
+import { HubConnection } from "@microsoft/signalr";
+import { createContext } from "react";
+
+export const ConnectionContext = createContext<HubConnection | null>(null);
