+++ frontend/src/app/page.tsx
@@ -1,225 +1,36 @@
 "use client";
 
-import GameScreen from "@/components/GameScreen";
-import JoinGameScreen from "@/components/JoinGameScreen";
-import LobbyScreen from "@/components/LobbyScreen";
-import MainMenu from "@/components/MainMenu";
-import ResultsScreen from "@/components/ResultsScreen";
-import { GameMode, GameState, Player } from "@/types/game";
-import { useMemo, useReducer, useState } from "react";
+import { useEffect, useState } from "react";
+import { ConnectionContext } from "./connectionContext";
+import Wrapper from "./wrapper";
+import { HubConnection, HubConnectionBuilder } from "@microsoft/signalr";
 import { useSearchParams } from "next/navigation";
-import * as signalR from "@microsoft/signalr";
 
 export default function Page() {
-  // useEffect(() => {
-  //   const connection = new signalR.HubConnectionBuilder()
-  //     .withUrl("http://localhost:5103/hub")
-  //     .build();
-  //   connection.start();
+  const searchParams = useSearchParams();
+  const isJoining = searchParams.get("join") ? true : false;
+  const gameId = searchParams.get("join") ?? crypto.randomUUID().toString();
 
-  //   connection.on("messageReceived", (username: string, message: string) => {
-  //     console.log(username, message);
-  //   });
-  // }, []);
-  const urlSearchParams = useSearchParams();
-  const isJoining = urlSearchParams.get("join") !== null;
-  const gameId = useMemo(
-    () => urlSearchParams.get("join") ?? crypto.randomUUID().toString(),
-    [],
+  const [conn, setConn] = useState<HubConnection | null>(null);
+
+  useEffect(() => {
+    const connection = new HubConnectionBuilder()
+      .withUrl("http://localhost:5103/hub")
+      .build();
+
+    connection
+      .start()
+      .then(() => setConn(connection))
+      .catch(() => setConn(null));
+  }, []);
+
+  return !conn ? (
+    <div className="flex h-full w-full items-center justify-center">
+      Making Connection...
+    </div>
+  ) : (
+    <ConnectionContext.Provider value={conn}>
+      <Wrapper gameId={gameId} isJoining={isJoining} />
+    </ConnectionContext.Provider>
   );
-
-  const currentPlayer: Player = {
-    id: 1,
-    name: "Player",
-    score: 0,
-    isHost: false,
-    progress: 0,
-  };
-
-  const [screen, setScreen] = useState<GameState>(isJoining ? "lobby" : "menu");
-  const [gameOps, dispatch] = useReducer(gameOpsreducer, {
-    gameId: gameId,
-    currentPlayer,
-    players: [],
-    gameMode: { type: "time", seconds: 10 },
-  });
-
-  return (
-    <main className="flex h-full w-full items-center justify-center">
-      {(() => {
-        switch (screen) {
-          case "menu":
-            return (
-              <MainMenu
-                onSelectMode={(gameMode) =>
-                  dispatch({ type: "setGameMode", gameMode })
-                }
-                onJoinGame={() => setScreen("joining")}
-                onCreateGame={() => {
-                  dispatch({ type: "createGame" });
-                  setScreen("lobby");
-                }}
-                onStartSinglePlayer={() => {
-                  setScreen("playing");
-                }}
-              />
-            );
-          case "joining":
-            return (
-              <JoinGameScreen
-                onJoinGame={() => {}}
-                onBackToMenu={() => setScreen("menu")}
-              />
-            );
-          case "lobby":
-            return (
-              <LobbyScreen
-                dispatch={dispatch}
-                onBackToMenu={() => {
-                  dispatch({ type: "exitLobby" });
-                  setScreen("menu");
-
-                  const connection = new signalR.HubConnectionBuilder()
-                    .withUrl("http://localhost:5103/hub")
-                    .build();
-
-                  connection
-                    .start()
-                    .then(() =>
-                      connection.send(
-                        "RemovePlayer",
-                        gameId,
-                        gameOps.currentPlayer.id,
-                      ),
-                    )
-                    .then(() => connection.stop())
-                    .catch();
-                }}
-                onStartGame={() => setScreen("playing")}
-                players={gameOps.players}
-                gameId={gameId}
-                currentPlayer={gameOps.currentPlayer}
-                selectedMode={gameOps.gameMode}
-              />
-            );
-          case "playing":
-            return (
-              <GameScreen
-                {...gameOps}
-                onGameEnd={() => setScreen("results")}
-                equations={[]}
-              />
-            );
-          case "results":
-            return (
-              <ResultsScreen
-                players={gameOps.players}
-                gameMode={gameOps.gameMode}
-                onBackToMenu={() => setScreen("menu")}
-                onPlayAgain={() => setScreen("playing")}
-              />
-            );
-        }
-      })()}
-    </main>
-  );
-}
-
-export type GameOps = {
-  currentPlayer: Player;
-  players: Player[];
-  gameMode: GameMode;
-  gameId: string;
-};
-
-export type GameOpsAction =
-  | {
-      type: "addPlayer";
-      player: Player;
-    }
-  | {
-      type: "setGameMode";
-      gameMode: GameMode;
-    }
-  | {
-      type: "createGame";
-    }
-  | {
-      type: "exitLobby";
-    }
-  | {
-      type: "nameChange";
-      name: string;
-    }
-  | {
-      type: "setPlayers";
-      players: Player[];
-    }
-  | {
-      type: "setCurrentPlayer";
-      player: Player;
-    };
-
-function gameOpsreducer(state: GameOps, action: GameOpsAction): GameOps {
-  switch (action.type) {
-    case "setCurrentPlayer":
-      return {
-        ...state,
-        currentPlayer: action.player,
-      };
-    case "addPlayer":
-      return {
-        ...state,
-        players: [...state.players, action.player],
-      };
-    case "setPlayers":
-      return {
-        ...state,
-        players: action.players,
-      };
-    case "setGameMode":
-      return {
-        ...state,
-        gameMode: action.gameMode,
-      };
-    case "createGame":
-      return {
-        ...state,
-        currentPlayer: {
-          ...state.currentPlayer,
-          isHost: true,
-        },
-      };
-    case "exitLobby":
-      return {
-        ...state,
-        players: state.players.map((p) => {
-          return {
-            ...p,
-            isHost: false,
-          };
-        }),
-        currentPlayer: {
-          ...state.currentPlayer,
-          isHost: false,
-        },
-      };
-
-    case "nameChange":
-      return {
-        ...state,
-        currentPlayer: {
-          ...state.currentPlayer,
-          name: action.name,
-        },
-        players: state.players.map((p) =>
-          p.id === state.currentPlayer.id
-            ? {
-                ...state.currentPlayer,
-                name: action.name,
-              }
-            : p,
-        ),
-      };
-  }
 }
